Update patch set 2
Patch Set 2:

> What exactly does require this?

I'm staging it here Alex, as far as I know this was introduced in 3.10 and it's present in 8992, 8994 and 8996 kernel.
The 8974 kernel doesn't have any reference for this API in its BT code and probably will fail if there's no workaround to support those older BT devices.

Patch-set: 2